New ver with tile based movement

+++ game/src/camera.rs
@@ -5,7 +5,7 @@ use super::common::Position;
 
 pub fn update(game: &mut Game) {
     let player_postion = game.world.get_component_from_entity_mut::<Position>(game.player).unwrap().as_mut().unwrap().value;    
-    let camera_x = (player_postion.x / ROOM_PIXEL_WIDTH as f32).floor() * -4.0;
-    let camera_y = (player_postion.y / ROOM_PIXEL_HEIGHT as f32).floor() * -4.0;
+    let camera_x = (player_postion.x as f32 / ROOM_PIXEL_WIDTH as f32).floor() * -4.0;
+    let camera_y = (player_postion.y as f32 / ROOM_PIXEL_HEIGHT as f32).floor() * -4.0;
     game.chroma.update_camera(camera_x, camera_y);
 }
